+++ packages/ts-newskit/src/components/coloured-text/index.tsx
@@ -1,0 +1,5 @@
+import { styled, getColorCssFromTheme, TextBlock } from 'newskit';
+
+export const ColouredText = styled(TextBlock)<{ $color?: string }>`
+  ${({ $color }) => $color && getColorCssFromTheme('color', $color)};
+`;