+++ packages/ts-newskit/src/components/navigation/secondary-menu/__tests__/secondary-desktop.test.tsx
@@ -12,6 +12,8 @@ const options = {
   isSelected: 'Home'
 };
 
+const mockClickHandler = jest.fn();
+
 describe('Secondary Menu Desktop', () => {
   afterEach(() => {
     jest.clearAllMocks();
@@ -20,27 +22,43 @@ describe('Secondary Menu Desktop', () => {
 
   it('should render snapshot', () => {
     const { asFragment } = render(
-      <SecondaryNavDesktop data={mainMenuItems} options={options} />
+      <SecondaryNavDesktop
+        data={mainMenuItems}
+        options={options}
+        clickHandler={mockClickHandler}
+      />
     );
     expect(asFragment()).toMatchSnapshot();
   });
   it('should render the menu item', () => {
     const { getByText } = render(
-      <SecondaryNavDesktop data={mainMenuItems} options={options} />
+      <SecondaryNavDesktop
+        data={mainMenuItems}
+        options={options}
+        clickHandler={mockClickHandler}
+      />
     );
     const title = getByText('Home');
     expect(title).toBeInTheDocument();
   });
   it('items should have ancher with href', () => {
     const { getAllByTestId } = render(
-      <SecondaryNavDesktop data={mainMenuItems} options={options} />
+      <SecondaryNavDesktop
+        data={mainMenuItems}
+        options={options}
+        clickHandler={mockClickHandler}
+      />
     );
     const title = getAllByTestId('buttonLink')[0];
     expect(title).toHaveAttribute('href', '/home');
   });
   it('should call handleSelect when clicked', () => {
     const { getAllByTestId } = render(
-      <SecondaryNavDesktop data={mainMenuItems} options={options} />
+      <SecondaryNavDesktop
+        data={mainMenuItems}
+        options={options}
+        clickHandler={mockClickHandler}
+      />
     );
     const Anchor = getAllByTestId('buttonLink')[0];
     fireEvent.click(Anchor);
@@ -48,14 +66,22 @@ describe('Secondary Menu Desktop', () => {
   });
   it('should render navitems', () => {
     const { getAllByRole } = render(
-      <SecondaryNavDesktop data={mainMenuItems} options={options} />
+      <SecondaryNavDesktop
+        data={mainMenuItems}
+        options={options}
+        clickHandler={mockClickHandler}
+      />
     );
     const list = getAllByRole('listitem');
     expect(list.length).toEqual(9);
   });
   it('should change the width of menudivader when screen size change', () => {
     const { getByTestId } = render(
-      <SecondaryNavDesktop data={mainMenuItems} options={options} />
+      <SecondaryNavDesktop
+        data={mainMenuItems}
+        options={options}
+        clickHandler={mockClickHandler}
+      />
     );
     const hr = getByTestId('divider');
     expect(hr).toHaveStyle('width: 100%');