+++ packages/ts-newskit/src/components/navigation/secondary-menu/__tests__/secondary-nav-mobile.test.tsx
@@ -25,6 +25,7 @@ describe('Secondary Menu', () => {
         onClick={() => {
           // noop
         }}
+        clickHandler={jest.fn()}
       />
     );
     expect(asFragment()).toMatchSnapshot();