+++ packages/ts-newskit/src/components/puzzles/hero-banner/styles.ts
@@ -45,7 +45,7 @@ export const HeroBannerContainer = styled(Stack)<StackProps>`
   position: relative;
   width: 100%;
   overflow: hidden;
-  ${getColorCssFromTheme('backgroundColor', 'puzzles040')};
+  ${getColorCssFromTheme('backgroundColor', 'sectionBrand040')};
   ${getMediaQueryFromTheme('xs')} {
     ${getSpacingCssFromTheme('paddingBlockStart', 'space100')};
     ${getSpacingCssFromTheme('paddingInline', 'space050')};
@@ -58,7 +58,7 @@ export const HeroBannerContainer = styled(Stack)<StackProps>`
 
 export const StyledStack = styled(Stack)`
   width: 100%;
-  ${getColorCssFromTheme('backgroundColor', 'puzzles040')};
+  ${getColorCssFromTheme('backgroundColor', 'sectionBrand040')};
 `;
 
 export const StyledHeroIconContainer = styled(Block)`