+++ packages/ts-newskit/src/slices/content-bucket-2/index.tsx
@@ -15,9 +15,7 @@ export const ContentBucket2 = ({
   articles,
   clickHandler
 }: ContentBucket2Props) => {
-  const [currentBreakpoint, setBreakpoint] = useState<BreakpointKeys | null>(
-    null
-  );
+  const [currentBreakpoint, setBreakpoint] = useState<BreakpointKeys>('xl');
   const breakpointKey = useBreakpointKey();
   useEffect(
     () => {
@@ -25,10 +23,6 @@ export const ContentBucket2 = ({
     },
     [breakpointKey]
   );
-
-  if (!currentBreakpoint) {
-    return null;
-  }
 
   const isMob = currentBreakpoint === 'xs' || currentBreakpoint === 'sm';