+++ packages/ssr/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.48.85](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.48.84...@times-components/ssr@2.48.85) (2023-10-03)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
+## [2.48.84](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.48.83...@times-components/ssr@2.48.84) (2023-10-02)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
 ## [2.48.83](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.48.82...@times-components/ssr@2.48.83) (2023-09-28)
 
 **Note:** Version bump only for package @times-components/ssr