+++ packages/ts-newskit/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.103.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.102.0...@times-components/ts-newskit@1.103.0) (2023-10-03)
+
+
+### Features
+
+* **TMRX-1513:** add component for handling coloured text ([#3562](https://github.com/newsuk/times-components/issues/3562)) ([b6a23c6](https://github.com/newsuk/times-components/commit/b6a23c62f04904b891ebe33ab54b3b1a8ffcc8a3))
+
+
+
+
+
+# [1.102.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.101.0...@times-components/ts-newskit@1.102.0) (2023-10-02)
+
+
+### Features
+
+* **TMRX-1547:** Add tracking to secondary navigation ([#3559](https://github.com/newsuk/times-components/issues/3559)) ([3d256fd](https://github.com/newsuk/times-components/commit/3d256fdd5caa2322e4a4f4a26eb07b0cb178d005))
+
+
+
+
+
 # [1.101.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.100.0...@times-components/ts-newskit@1.101.0) (2023-09-28)